Made eda_viz.py, moved method that gets dataframe from date range to its own py file
